+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.37.0] 2021-10-26
+
+- Added Alert component
+- Fixed LabelSwitch width
+- Added 'label' tag option to Typography component
+
 ## [1.36.3] - 2021-10-21
 
 - LabelSwitch no longer breaks line on larger labels